Add methods for team and note management in HackmdClient and related classes

+++ hackmd-client-core/src/main/java/io/github/yuokada/hackmd/core/Team.java
@@ -10,4 +10,4 @@ public record Team(
     String description,
     TeamVisibility visibility,
     boolean upgraded,
-    double createdAt) {}
+    long createdAt) {}
